refactor: stack overflow track quota
