+++ packages/socials/src/stack-overflow/get-user/index.ts
@@ -1,4 +1,4 @@
-import { trackQuota } from "../track-quota";
+import { trackQuota } from "../response/track-quota";
 import { getStackOverflowUserResponseSchema } from "./schema";
 
 async function getStackOverflowUserData() {
